+++ common/include/ag_version.h
@@ -1,0 +1,2 @@
+#pragma once
+#define AG_DNSLIBS_VERSION "1.5.0"